Basic introductory app for League of Legends players
This app is very basic. It allows players to select which champion they're going to be using, then either hit 'enter' or click on the submit button to open up the proper na.op.gg webpage dedicated to their champions. This page will have statistical analysis and provide the user with the best runes and item builds to succeed with their champion.